requirements.txt
ansicon==1.89.0
asgiref==3.7.2
asttokens==2.4.1
blessed==1.19.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.14.0
fire==0.4.0
halo==0.0.31
idna==3.2
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
Jinja2==3.1.3
jinxed==1.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
log-symbols==0.0.14
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.9
npm==0.1.1
numpy==1.26.3
openpyxl==3.1.2
optional-django==0.1.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathspec==0.9.0
pillow==10.3.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.10.0
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.3
pywin32==306
pyzmq==25.1.2
readchar==2.0.1
regex==2023.10.3
requests==2.26.0
requests-toolbelt==0.9.1
runflare==1.1.8
setuptools==69.0.3
six==1.16.0
spinners==0.0.24
sqlparse==0.4.4
stack-data==0.6.3
termcolor==1.1.0
tornado==6.4
traitlets==5.14.1
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.7
virtualenv==20.26.2
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
websockets==10.0
wheel==0.42.0
whitenoise==6.7.0
